(feat) Atualizando método de baixa de resíduos para usar ResiduoDownturnRequestDTO e ajustando tipo de quantidade para Long

+++ src/main/java/org/purpura/apimg/dto/schemas/empresa/residuo/Downturn.java
@@ -23,5 +23,5 @@ public class Downturn {
     @NotNull(
             message = "A quantidade de baixa deve ser informada e não nula"
     )
-    private Double quantidade;
+    private Long quantidade;
 }
